# Conflicts:
#   server/src/routes/image.js
#   server/src/routes/userIssue.js

+++ server/src/routes/feedback.js
@@ -1,4 +1,6 @@
-import {Feedback} from "../models";
+import {Feedback, Issue, User} from "../models";
+import {tokenManager} from "../tokenManager";
+import * as passwordHash from "../passwordHash";
 require('dotenv').config();
 
 type Request = express$Request;
@@ -6,14 +8,20 @@ type Response = express$Response;
 
 const app = require('../app');
 //POST one feedback
-app.post('/secure/feedback', (req: Request, res: Response) => {
+app.post('/feedback', (req: Request, res: Response) => {
     if (!(req.body instanceof Object)) return res.sendStatus(400);
-    return Feedback.create({
-        name: req.body.name,
-        content: req.body.content,
-        issueId: req.body.issueId,
-        userId: req.body.userId
-    }).then(count => (count ? res.sendStatus(200) : res.sendStatus(404)));
+
+    let tokenData = tokenManager.verifyToken(req.headers['x-access-token']);
+    if (tokenData) {
+        return Feedback.create({
+            name: req.body.name,
+            content: req.body.content,
+            issueId: req.body.issueId,
+            userId: req.body.userId
+        }).then(count => (count ? res.sendStatus(200) : res.sendStatus(404)));
+    } else {
+        res.sendStatus(401);
+    }
 });
 
 //GET all feedback for one issue, ordered ascending
@@ -27,6 +35,7 @@ app.get('/issues/:id/feedback/:lim/limit/:off/offset/asc', (req: Request, res: R
         order: [
             ['createdAt']
         ]
+
         }).then(
         issue => (issue ? res.send(issue) : res.sendStatus(404))
     );
@@ -56,11 +65,38 @@ app.get('/feedback/:id', (req: Request, res: Response) => {
         feedbak => (feedbak ? res.send(feedbak) : res.sendStatus(404))
     );
 });
+
+//UPDATE one feedback with id
+app.put('/feedback/:id', (req: Request, res: Response) => {
+    if (!(req.body instanceof Object)) return res.sendStatus(400);
+
+    let tokenData = tokenManager.verifyToken(req.headers['x-access-token']);
+    if (tokenData) {
+        return Feedback.update(
+            {
+                content: req.body.content
+            },
+            {
+                where: {
+                    feedbackId: req.params.id
+                }
+            }
+        ).then(feedback => (feedback ? res.sendStatus(200) : res.sendStatus(404)));
+    } else {
+        res.sendStatus(401);
+    }
+});
+
 //DELETE one feedback with id
 app.delete('/feedback/:id', (req: Request, res: Response) => {
-    return Feedback.destroy({
-        where: {
-            feedbackId: req.params.id
-        }
-    }).then(count => (count ? res.sendStatus(200) : res.sendStatus(404)));
+    let tokenData = tokenManager.verifyToken(req.headers['x-access-token']);
+    if (tokenData) {
+        return Feedback.destroy({
+            where: {
+                feedbackId: req.params.id
+            }
+        }).then(count => (count ? res.sendStatus(200) : res.sendStatus(404)));
+    } else {
+        res.sendStatus(401);
+    }
 });
