+++ server/src/routes/userIssue.js
@@ -58,15 +58,25 @@ app.get('/secure/UsersIssues/:rank/status/:statusId', (req: Request, res: Respon
     }).then(user => (user ? res.send(user) : res.sendStatus(404)));
 });
 
-app.post('/secure/users/:userId/issues/:issueId', (req: Request, res: Response) => {
-    return UserIssue.create({
-        userId: req.params.userId,
-        issueId: req.params.issueId
-    }).then(count => (count ? res.sendStatus(200) : res.sendStatus(404)));
+app.post('/users/:userId/issues/:issueId', (req: Request, res: Response) => {
+    let tokenData = tokenManager.verifyToken(req.headers['x-access-token']);
+    if (tokenData) {
+        return UserIssue.create({
+            userId: req.params.userId,
+            issueId: req.params.issueId
+        }).then(count => (count ? res.sendStatus(200) : res.sendStatus(404)));
+    } else {
+        res.sendStatus(401);
+    }
 });
 
-app.delete('/secure/users/:userId/issues/:issueId', (req: Request, res: Response) => {
-    return UserIssue.destroy({ where: { userId: req.params.userId, issueId: req.params.issueId } }).then(count =>
-        count ? res.sendStatus(200) : res.sendStatus(404)
-    );
+app.delete('/users/:userId/issues/:issueId', (req: Request, res: Response) => {
+    let tokenData = tokenManager.verifyToken(req.headers['x-access-token']);
+    if (tokenData) {
+        return UserIssue.destroy({ where: { userId: req.params.userId, issueId: req.params.issueId } }).then(count =>
+            count ? res.sendStatus(200) : res.sendStatus(404)
+        );
+    } else {
+        res.sendStatus(401);
+    }
 });
