+++ server/src/routes/userMunicipal.js
@@ -1,5 +1,6 @@
 //@flow
-import { Municipal, User, UserMunicipal } from '../models';
+import {Municipal, User, UserIssue, UserMunicipal} from '../models';
+import {tokenManager} from "../tokenManager";
 
 type Request = express$Request;
 type Response = express$Response;
@@ -7,7 +8,7 @@ type Response = express$Response;
 const app = require('../app');
 
 
-app.get('/secure/users/:id/mun', (req: Request, res: Response) => { //userMun
+app.get('/users/:id/mun', (req: Request, res: Response) => { //userMun
     return Municipal.findAll({
         include: [
             {
@@ -20,16 +21,27 @@ app.get('/secure/users/:id/mun', (req: Request, res: Response) => { //userMun
     }).then(user => (user ? res.send(user) : res.sendStatus(404)));
 });
 
-app.post('/secure/users/:userId/mun/:munId', (req: Request, res: Response) => {
+app.post('/users/:userId/mun/:munId', (req: Request, res: Response) => {
     if (!(req.body instanceof Object)) return res.sendStatus(400);
-    return UserMunicipal.create({
-        userId: req.params.userId,
-        munId: req.params.munId
-    }).then(count => (count ? res.sendStatus(200) : res.sendStatus(404)));
+
+    let tokenData = tokenManager.verifyToken(req.headers['x-access-token']);
+    if (tokenData) {
+        return UserMunicipal.create({
+            userId: req.params.userId,
+            munId: req.params.munId
+        }).then(count => (count ? res.sendStatus(200) : res.sendStatus(404)));
+    } else {
+        res.sendStatus(401);
+    }
 });
 
-app.delete('/secure/users/:userId/mun/:munId', (req: Request, res: Response) => {
-    return UserMunicipal.destroy({ where: { userId: req.params.userId, munId: req.params.munId } }).then(count =>
-        count ? res.sendStatus(200) : res.sendStatus(404)
-    );
+app.delete('/users/:userId/mun/:munId', (req: Request, res: Response) => {
+    let tokenData = tokenManager.verifyToken(req.headers['x-access-token']);
+    if (tokenData) {
+        return UserMunicipal.destroy({ where: { userId: req.params.userId, munId: req.params.munId } }).then(count =>
+            count ? res.sendStatus(200) : res.sendStatus(404)
+        );
+    } else {
+        res.sendStatus(401);
+    }
 });
